divide complexed file, lint
