+++ src/index.ts
@@ -47,7 +47,7 @@
 /// <reference path="Renderer.ts" />
 /// <reference path="Surface.ts" />
 /// <reference path="Label.ts" />
-/// <reference path="BitmapFont.ts" />
+/// <reference path="Glyph.ts" />
 /// <reference path="FilledRect.ts" />
 /// <reference path="Pane.ts" />
 /// <reference path="SurfaceEffector.ts" />
@@ -65,12 +65,15 @@
 /// <reference path="AudioSystemManager.ts" />
 /// <reference path="CompositeOperation.ts" />
 /// <reference path="EventFilter.ts" />
+/// <reference path="GlyphFactory.ts" />
 /// <reference path="Font.ts" />
 /// <reference path="GameMainParameterObject.ts" />
 /// <reference path="LocalTickMode.ts" />
 /// <reference path="MultiLineLabel.ts" />
 /// <reference path="NinePatchSurfaceEffector.ts" />
 /// <reference path="PathUtil.ts" />
+/// <reference path="FontFamily.ts" />
+/// <reference path="BitmapFont.ts" />
 /// <reference path="SystemLabel.ts" />
 /// <reference path="TextAlign.ts" />
 /// <reference path="TickGenerationMode.ts" />
